+++ src/Tests.Worker/TimerWorkerItemSpecs/StartTests.cs
@@ -1,30 +1,11 @@
 using FakeItEasy;
-using FatCat.Worker;
-using FatCat.Worker.Wrappers;
 using FluentAssertions;
-using Humanizer;
 using Xunit;
 
 namespace Tests.FatCat.Worker.TimerWorkerItemSpecs;
 
-public class StartTests
+public class StartTests : TimerWorkerItemTests
 {
-	private readonly TimerWorkerItem timerWorkerItem;
-	private TimeSpan interval;
-	private int numberOfTimesToRun;
-	private ITimerWrapper timerWrapper;
-	private ITimerWrapperFactory timerWrapperFactory;
-	private bool waitForDelay;
-	private IWorkerItem workerItem;
-
-	public StartTests()
-	{
-		SetUpTimerWrapperFactory();
-		SetUpWorkerItem();
-
-		timerWorkerItem = new TimerWorkerItem(timerWrapperFactory);
-	}
-
 	[Fact]
 	public void CreateTimerWrapper()
 	{
@@ -81,33 +62,5 @@ public class StartTests
 
 		A.CallTo(() => timerWrapper.Start())
 		.MustHaveHappened();
-	}
-
-	private void SetUpTimerWrapperFactory()
-	{
-		timerWrapperFactory = A.Fake<ITimerWrapperFactory>();
-
-		timerWrapper = A.Fake<ITimerWrapper>();
-
-		A.CallTo(() => timerWrapperFactory.CreateTimerWrapper())
-		.Returns(timerWrapper);
-	}
-
-	private void SetUpWorkerItem()
-	{
-		workerItem = A.Fake<IWorkerItem>();
-
-		interval = 45.Seconds();
-		numberOfTimesToRun = -1;
-		waitForDelay = true;
-
-		A.CallTo(() => workerItem.Interval)
-		.ReturnsLazily(() => interval);
-
-		A.CallTo(() => workerItem.WaitOnWorkBeforeDelay())
-		.ReturnsLazily(() => waitForDelay);
-
-		A.CallTo(() => workerItem.NumberOfTimesToRun())
-		.ReturnsLazily(() => numberOfTimesToRun);
 	}
 }
